Fix to get double value from args when 'seekTo' occurs

+++ android/src/main/java/com/amazonaws/ivs/reactnative/player/AmazonIvsViewManager.kt
@@ -44,7 +44,7 @@ class AmazonIvsViewManager : SimpleViewManager<AmazonIvsView>() {
       Commands.PAUSE.ordinal -> view.pause()
       Commands.TOGGLE_PIP.ordinal -> view.togglePip()
       Commands.SEEK_TO.ordinal -> {
-        args?.getInt(0)?.let { position ->
+        args?.getDouble(0)?.let { position ->
           view.seekTo(position.toLong())
         }
       }
